Requested Changes v2

+++ minecraft/math/BoundingBox.h
@@ -8,12 +8,12 @@ class BoundingBox {
 
 public:  
     
-    float minX;
-    float minY;
-    float minZ;
-    float maxX;
-    float maxY;
-    float maxZ;
+    int minX;
+    int minY;
+    int minZ;
+    int maxX;
+    int maxY;
+    int maxZ;
 
     void orientBox(int, int, int, int, int, int, int, int, int, int);
     std::unique_ptr<IntArrayTag> createTag() const;
